task-helper/cloud-function/requirements.txt
functions-framework==3.*
google-generativeai
requests